app/requirements.txt
streamlit
pillow==11.0.0
folium==0.18.0 
streamlit-folium==0.23.1
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
